Count stock at its cost when it sells, in fractions, and by location
Profit counted every purchase of stock as an expense the day it was bought, so a month of
restocking read as a loss and the month the stock sold read as almost pure profit. A product that
tracks inventory now has one stock count on both the purchase and sale forms, buying it adds to
stock rather than expenses, and selling it fixes what that stock cost on the sale line. Stock
already on hand was expensed when bought, so it becomes opening units that sell at no cost, and
editing a transaction saved before this keeps its old treatment.

Stock quantities hold decimals, so 2.5 kg no longer rounds to 3. That moves the file format to
version 3, so an older build says it is out of date rather than failing on a fractional quantity.

Stock moves between locations with a transfer, and a sale or purchase line asks which location
only when the product is stocked at more than one. A stock record created by a sale now takes the
product's cost price rather than the selling price. The Stock Adjustments totals are decimal too.

The rules are in Calculations.md section 14.

+++ ArgoBooks.Core/Services/SpreadsheetImportService.cs
@@ -3833,9 +3833,9 @@ Respond with ONLY a JSON array, one entry per product in the same order:
 
             // Handle Reorder Point and Overstock Threshold
             if (Set("Reorder Point"))
-                product.ReorderPoint = GetInt(row, headers, "Reorder Point");
+                product.ReorderPoint = GetDecimal(row, headers, "Reorder Point");
             if (Set("Overstock Threshold"))
-                product.OverstockThreshold = GetInt(row, headers, "Overstock Threshold");
+                product.OverstockThreshold = GetDecimal(row, headers, "Overstock Threshold");
 
             // Set TrackInventory based on whether reorder/overstock values are set
             if (product.ReorderPoint > 0 || product.OverstockThreshold > 0)
@@ -3887,11 +3887,11 @@ Respond with ONLY a JSON array, one entry per product in the same order:
             if (Set("Location ID"))
                 item.LocationId = locationId;
             if (Set("In Stock"))
-                item.InStock = GetInt(row, headers, "In Stock");
+                item.InStock = GetDecimal(row, headers, "In Stock");
             if (Set("Reserved"))
-                item.Reserved = GetInt(row, headers, "Reserved");
+                item.Reserved = GetDecimal(row, headers, "Reserved");
             if (Set("Reorder Point"))
-                item.ReorderPoint = GetInt(row, headers, "Reorder Point");
+                item.ReorderPoint = GetDecimal(row, headers, "Reorder Point");
             if (Set("Unit Cost"))
                 item.UnitCost = GetDecimal(row, headers, "Unit Cost");
             if (Set("Last Updated"))
@@ -4518,7 +4518,7 @@ Respond with ONLY a JSON array, one entry per product in the same order:
                             {
                                 Id = $"INV-ITM-{nextNum:D3}",
                                 ProductId = productId,
-                                InStock = GetInt(row, headers, "Total Qty")
+                                InStock = GetDecimal(row, headers, "Total Qty")
                             };
                             data.Inventory.Add(newInv);
                             inventoryItemId = newInv.Id;
@@ -4830,11 +4830,11 @@ Respond with ONLY a JSON array, one entry per product in the same order:
             if (Set("Type"))
                 adjustment.AdjustmentType = ParseEnum(GetString(row, headers, "Type"), AdjustmentType.Set);
             if (Set("Quantity"))
-                adjustment.Quantity = GetInt(row, headers, "Quantity");
+                adjustment.Quantity = GetDecimal(row, headers, "Quantity");
             if (Set("Previous Stock"))
-                adjustment.PreviousStock = GetInt(row, headers, "Previous Stock");
+                adjustment.PreviousStock = GetDecimal(row, headers, "Previous Stock");
             if (Set("New Stock"))
-                adjustment.NewStock = GetInt(row, headers, "New Stock");
+                adjustment.NewStock = GetDecimal(row, headers, "New Stock");
             if (Set("Reason"))
                 adjustment.Reason = GetString(row, headers, "Reason");
             if (Set("Reference Number"))
@@ -4999,9 +4999,9 @@ Respond with ONLY a JSON array, one entry per product in the same order:
             var lineItem = new PurchaseOrderLineItem
             {
                 ProductId = GetString(row, headers, "Product ID"),
-                Quantity = GetInt(row, headers, "Quantity"),
+                Quantity = GetDecimal(row, headers, "Quantity"),
                 UnitCost = GetDecimal(row, headers, "Unit Cost"),
-                QuantityReceived = GetInt(row, headers, "Quantity Received")
+                QuantityReceived = GetDecimal(row, headers, "Quantity Received")
             };
 
             if (!lineItemsByPo.ContainsKey(poId))
